Adds appsembler roles and inventory
We require our own roles repo for the playbook to run.
Especially if we're running "ansible-playbook --tags=letsencrypt:run ..."
in an automated fashion.

We also need a list of edx app nodes so we copy over a inventory file
called "app_nodes_inventory" that lives in the "edx_ansible_app_dir".

Putting that all together the calling command should be:

"cd /edx/app/edx_ansible/edx_ansible/playbooks"
"../../venvs/edx_ansible/bin/ansible-playbook -i ../../app_nodes_inventory --tags=letsencrypt:run
amc.yml"

Or something similar depending on what we decide is best but we need to
take care to invoke the correct venv and to be careful with the paths.