add active monitoring of backups
With our current setup, we get an exception in Sentry if the backup
script encounters an error. However, it could still fail silently and
we would have no way of knowing until we go to restore and find that
the backup file is missing.

This is a first step in addressing that gap. When this one runs, it
checks the most recent backup file in the Cloud Storage bucket (this
first pass only supports Google Cloud for now) and raises an alert if
it is more than 25 hours old (the assumption is that we are doing
nightly backups).

Really, this is still just a placeholder until we get a comprehensive
monitoring and alerting system in place. At that point, instead of
raising an exception here, this script will just submit the 'hours
since last backup' value as a metric to the monitoring system and we
can alert when that crosses a threshold. That would also allow us to
also catch a situation where this monitoring script doesn't run (by
alerting on data missing from the metric).