push backups monitoring timestamp to prometheus pushgateway
As mentioned in #135, monitoring backups by directly creating sentry
alerts when the age exceeded a threshold was a stopgap measure while
we got some more monitoring infrastructure in place.

I have prometheus running now so this changes the backup monitoring to
just get the timestamp of the most recent backup file and pushes that
to the prometheus pushgateway.

From there, we can set Prometheus to alert when `time() -
backups_timestamp` exceeds a threshold. That lets manage the
thresholds a little better (ie, at the alerting end, not setting it in
this script) so we can change them later on or set different ones for
different projects while this script stays the same. Since it's all
metric data going into prometheus, it also us generate Grafana
dashboards to show the current state of the backup monitoring.

Since I don't expect many of you to be too familiar with Prometheus, a
couple quick notes:

* Prometheus normally "pulls" data by actively scraping
  endpoints. That doesn't really work for batch jobs like this, so
  there's a "push gateway" that scripts like this can push data into
  via HTTP and Prometheus is set to then pull from there. See:
	https://github.com/prometheus/pushgateway
* There's a nice python client for the push gateway, but I've just
  done a simple POST request via `requests` here so we don't need to
  install an additional python library on every server. (I *think*
  it's safe to assume that `requests` will be available)

I've left in some of the sentry functionality to handle cases where
this script runs into an error or the backups are not there at all. In
theory, those should be redundant since Prometheus can alert on
missing metrics, but they ought to help us catch a problem a little
faster and provide a little more context for debugging.