run collectstatic against a temp dir
to avoid breaking production while it runs

The basic approach:

* make a temp directory
* generate a new 'override' django config that imports the existing
  config and then just sets `STATIC_ROOT` to point to that temp
  directory
* runs the paver collectstatic command using that override config
* mv the old static root away
* mv the temp directory in place of the old static root
* clean up the override config

Now, the only point at which the running app could have problems with
the static files is for the brief period between the two moves. It's
not perfect, but that's a much smaller window than the 10-15 minutes
it took for the collectstatic to run before.