skip the AWS role for Tahoe
Tahoe isn't deployed on AWS, so it's not relevant. I also ran into an
error even with the `cloud_provider` set to "gcp". The `when` clause
on one of the tasks still expects to have a bunch of other stuff
defined and causes problems, so having this role included also
requires a bunch of additional useless config.

This role was removed in the `ficus` branch and now I see why.