remove hash from staticfiles override
Hawthorn no longer uses the `EDX_PLATFORM_REVISION` hash in the
CMS `STATIC_ROOT` (see: https://github.com/edx/edx-platform/pull/16819)
instead just using `/studio` for the CMS.

That means that when we override the config (for the zero-downtime
deploys), we need to match that change.

This should fix the problems we are having with hawthorn deploys
breaking and requiring a manual `update-static-assets` every time so
it doesn't break on the missing `webpack-stats.json`.