install rabbitmq earlier in playbook
Since edX actually depends on it for Celery, it should be installed
earlier.

My experience is that when it's installed after, `edxapp`, etc. can
succeed in installing, but the lms/cms processes will be stuck in a
loop failing to start because they can't connect to RabbitMQ. If the
whole playbook succeeds, that's usually only for a few minutes and not
very noticable on a new deploy. But if something else fails in between
them, we end up with thousands of Sentry events piling up.

Simple fix is to install RabbitMQ earlier in the process before we
start on any of the roles that might need it.