install fluentd stackdriver setup for hawthorn pro
We use this role on Tahoe already. Default configuration ingests most
of the basic edX/nginx/etc logs and passes them to the Stackdriver
Logging service for the project.

This more or less replaces the logstash/elasticsearch setup that we
used to use but shut off because it was under-utilized and
buggy (removing those roles is something we should also think about).

Here's an internal writeup I did on the setup: https://appsembler-infrastructure.appspot.com/post/stackdriver-logging/