install rabbitmq earlier in the deploy for hawthorn pro
With the existing setup, rabbitmq isn't installed until well after
`edxapp`. That means that on a new install, when `edxapp` sets up and
starts celery, it has no rabbitmq to talk to and we get hundreds of
Sentry errors as the celery worker dies and is restarted by supervisor
over and over again until rabbitmq is finally installed (which might
be much later if we run into other issues to debug in between them).

This just moves `rabbitmq` up to a much earlier part of the playbook
so it is installed around the same time as `memcached`,
`elasticsearch`, and other basic services that `edxapp` depends on
rather than afterwards. I don't see anything in the `rabbitmq` role
that would make it dependent on any of the earlier ones.