override nginx systemd to retry restarts
On Tahoe, nginx depends on the GCS Fuse mount to be available. When a
machine boots, that mount takes a few seconds to come up. If nginx
tries to start before that, it fails to read SSL certs from the mount
and fails. The default systemd config sets `Restart=no`, so it just
doesn't try again and we have to manually start it.

This should get systemd to keep trying to start nginx even if it fails
at first. This should help minimize downtime when rebooting a machine.