limit celerybeat to one server
In a deployment with more than one edxapp server, we need to make sure
that `lms_celerybeat` and `cms_celerybeat` are only installed on one
of them; otherwise we get duplicate scheduled tasks, conflicts, and
general pain.

This is about the best solution I could come up with using
ansible. Ansible includes a `run_once` directive, but that has the
problem that it will always run it once even if one of the hosts is
temporarily unreachable during the deploy. Ie, if the one with celery
beat is temporarily down during a deploy, ansible will
install/configure it on one of the other hosts and later when the
initial one comes back, they will both have it. The
`inventory_hostname == ansible_play_hosts_all[0]` basically tells it
to install it on the first host in the list as it's configured. As
long as we don't change the order of entries in the `inventory`, it
should be fine.