only use the request.path field for sampling decisions
It turns out that beeline sends spans independently and reconstructs
them into the full trace on their end. That means that the sampling
decision is made only on the fields in the span. On spans that aren't
the "root" span of a request, it doesn't have the full context.

In our case, that meant that the spans corresponding to traced
middleware, functions, etc. `request.path` and `respons.status_code`
are not available to the sampler. So it would always send them. Only
the root spans have those fields.

That means that downsampling `/heartbeat` requests is only affecting
those root spans. We currently have about five spans generated per
request, and four of them are not downsampled, so it's not having much
of an effect and we're still blowing past our quotas.

The "correct" solution to this problem is to run samproxy:
https://github.com/honeycombio/samproxy which reconstructs the traces
and does sampling logic on the whole thing before sending along to
Honeycomb. We might do that eventually, but it's a lot and I'd rather
not set that up yet.

My compromise is that I've used `beeline.add_trace_field(...)` to
propagate the `request.path` to child spans in
https://github.com/appsembler/edx-platform/pull/626

So the sampler should now have that one available. However,
`response.status_code` isn't known until the request is done, so
unless we time travel, we can't add that to child spans that beeline
is sampling before the request completes.

We can sample on `request.path` and it should now affect all of the
child spans as well. The catch is that we can't use the response code
in our sampling decision.