honeycomb: downsample `/event` requests
With our recent traffic increase, we're coming up on our honeycomb
quota again. I went through our data and noticed that about half the
requests we get (that aren't for `/heartbeat`, which we already
downsample) are for `/event`. My understanding is that the front-end
makes a background `POST` to that on basically every page load and all
it does is generate a tracking log/segmentIO event. We don't have any
interesting instrumentation on those requests and they aren't in the
critical path of a user's experience, so they are a good candidate to
also downsample to keep within our honeycomb quota.