use `hawthorn` tag for docker image
`latest` is going to start to be a problem as we work on the juniper
upgrade. Building and pushing with a `hawthorn` tag makes it clear how
we can do the same for juniper and work in parallel.

Devstack's `docker-compose.yml` will need to be updated to pull from
the `hawthorn` tag instead of `latest` as well.