downgrade urllib3
The backups script for GCP installs a bunch of python libraries
system-wide. One of those as a side-effect now results in a version of
`urllib3` getting installed system-wide that then breaks ansible on
subsequent deploys:

https://appsembler.atlassian.net/wiki/spaces/ED/pages/390004931/Failed+to+install+repo+key

A proper fix would be to move the whole backups script setup into a
virtualenv so we don't get conflicts. But since this is only a problem
on Hawthorn (Juniper uses a newer version of Ansible that is OK with
upgrading urllib3), the simple workaround for now is to just downgrade
urllib3 back to a known safe version at the end of the task.