run node exporter role earlier in playbook
since `edxapp` writes git info out to the node exporter textfile
collector directory (see `edxapp/tasks/git_info.yml`), it needs that
directory to exist before it can run. On a first install, the `edxapp`
role runs before `node_exporter`, so that breaks.

This moves the node exporter role up earlier in the playbook so it has
run and created the directory before any of the other roles that might
expect it to already exist.

(this has already been fixed in Tahoe and Juniper, so just backporting
to hawthorn)