python3 fix for lms template jinja
`iteritems()` was removed from python3. This causes the ansible
playbook to fail when it processes this template.

```
{"ansible_loop_var": "item", "changed": false, "item":
"maxitest5-staging.appsembler.com", "msg": "AnsibleUndefinedVariable:
'dict object' has no attribute 'iteritems'"}
```

Upstream fixed it in `lms.j2` and `cms.j2` in 2d4c3d0683 but since we
have a slightly refactored setup with `lms-common-settings.j2` split
out, we didn't get the fix and have to do this manually in our fork.