juniper prod deploy tweaks
Hit a few issues during the fresh install of juniper prod:

* ansible needed `python3-pip` and `python3-mysqldb` packages (I'm not
sure why prod seems to be using `python3` for ansible stuff and that
wasn't an issue on staging...)
* update forum version to juniper (staging will need a re-deploy; I
think it's still using the hawthorn forum image)

I also ran into an issue where the first deploy failed because there's
still some weird circular dependency going on with `edx_ansible`. I
had to disable it, deploy to get other dependencies installed, then
re-enable it. After that it's not a problem, but eventually we need to
fix that.