only run static asset pipeline on one edxapp server
Since we're now deploying static files to S3, the whole slow static
files pipeline only needs to run once. We can run it on the first
edxapp server, and then skip it on the 2nd and 3rd.

*Currently* our standalone pro deployments all just have one edxapp
server, so this shouldn't affect them.

If we ever have standalone sites with multiple edxapp servers and
we *don't* configure S3 static files on those deployments, we'll need
to figure out another way to do this (but we'd probably recommend just
switching them to S3 instead).