Tahoe: synchronize webpack-stats.json to all edxapp servers
attempt at fixing the webpack inconsistency issue.

Ansible's `synchronize` and `delegate_to` combine in a bit of a
confusing way, but the result is that it copies the designated file
from the `delegate_to` machine to the current host in the play. With
the `when` conditional on there, this should only run on the hosts
that aren't the first in the rolling deploy (the first is the one that
generates the files in the first place).