recursively fix directory ownership of course data dir
make sure the course data dir is set up correctly.
sometimes engineers making manual changes on the server
run shell commands as `edxapp` instead of `www-data`
and that can leave directories with the wrong ownership
in place. We run this on every deploy as an additional
fallback in case they don't notice.