exclude the mysql_init (localhost) from the common roles
We want these roles applied to all hosts, but the Tahoe inventory
files have special entries like:

```
[mysql_init]
127.0.0.1
```

Targetting `all` causes it to want to run those on `127.0.0.1` as
well, which clearly doesn't work.

We could either directly include all the other groups with something
like:

```
hosts: edx,services,workers,forum
```

But then if we someday add another group (eg, if we merge the Metabase
setup into this playbook), we'd need to remember to include it there
as well.

Instead, we stick with `all`, but then explicitly exclude the one odd
one.

(a separate question is whether we even use that `mysql_init` group
for anything or if it's leftover from previous work)