reduce honeycomb usage by sampling "boring" requests
We're running up on our honeycomb quota since the Django 2 upgrade,
since we now get spans for every single mysql query. That's good for
improving debugging capabilities, but we need to reduce our usage to
keep costs down.

This adjusts our sampling approach to identify "boring" requests that
are unlikely to be involved in us trying to track down a tricky bug or
performance problem. Spans associated with those "boring" requests are
then aggressively downsampled.

As a starting point, I'm using "anything that has a 200 OK response
and completed in less than 100ms" as the working definition of
"boring". This is a bit of a guess and we'll probably adjust this over
time.

But based on the last 24 hours of data, this would knock out a
significant portion of them (though there's a lot of overlap with the
downsampling of `/heartbeat` requests, which we already do)