check lms/cms heartbeat before re-adding node to load balancer
On a Tahoe rolling deploy, require a successful request to the
`/heartbeat` endpoint of LMS and CMS.

In the case of a major runtime bug (eg, a bad import in theme code, or
an invalid database connection setting, etc), we want to detect the
problem as quickly as possible and fail the deploy quickly rather than
continuing to deploy and route traffic to a bad instance.

We can't rely on checking the status of the process in Supervisor
because it typically takes a good 30 seconds or so for the lms/cms to
start up. During that time, Supervisor just sees that the process is
running and assumes it's healthy. If it finishes starting up, then
immediately crashes, Supervisor restarts it but then goes right back
to thinking it's healthy while it's still starting up.

So instead, we use a more direct check and hit the `/heartbeat`
endpoint on each (via http://localhost:8000/heartbeat and
http://localhost:8010/heartbeat). If we actually get a successful
response from those, we know that the process is fully started, is
actually running at least well enough to respond to requests, and
should have a valid connection to the MySQL and MongoDB databases. If
all of those are true, we can be much more confident in turning nginx
back on and routing traffic to this instance again.

Since we do know that sometimes it takes a while for the lms/cms to
start, we try up to three times, waiting one minute between each.

On a bad deploy, this *should* mean that it will get to the first
edxapp instance, turn off nginx to remove it from the load balancer,
update the code/settings/etc, try to restart the processes, then hit
this check, which ought to fail (after three minutes of trying). The
whole deploy will then fail, hopefully loudly so developers notice. At
that point, we will be down one node, but the second and third edxapp
nodes should still be running the previous version of the code and
will be handling traffic. We should see that the deploy failed,
investigate on the first edxapp instance, come up with a fix, and
schedule another deploy.