use self-signed SSL certs for health check
Installing the `ssl-certs` package creates the
`ssl-cert-stakeoil.{pem,key}` files when it's installed. Use those for
our healthcheck config so it won't depend on our wildcard certs being
present (so we can eventually remove them once we're behind the HTTPS
LB, at which point we'll be using these self-signed certs for the rest
of the config as well).