add `nginx_use_gcp_https_lb` toggle
Defaults to `false`. When enabled, switches the main LMS/CMS nginx
config to use self-signed certs instead of the wildcard certs.

The original idea was to have a `ENABLE_CUSTOM_DOMAINS_NGINX_CONFIG`
toggle, but there's already `nginx_enable_custom_domains` which is
doing most of that.

This additional toggle is only needed because we also want to not
break things for the standalone deployments, which have
`nginx_enable_custom_domains` disabled, but should not be switched
over to self-signed certs (yet. I think we could make an argument that
we ought to move the standalone deployments to HTTPS LB eventually as
well).