tahoe nginx-only playbook
adds a playbook that only handles the nginx-related parts of
Tahoe (copied from `amc.yml` and pruned down to just the relevant
parts).

This should hopefully speed things up a bit when we test and roll out
the HTTPS LB.

Once we're fully moved to that, hopefully we won't really have a need
for this playbook anymore and we can delete it.